requirements.txt
python-telegram-bot==22.6
sqlalchemy
python-dotenv
requests
beautifulsoup4
